refactor(components): restructure component folders for modular organization
- page.tsx:
  - update import paths to match new component structure

- components/:
  - move AnimatedContainer.tsx → components/ui/AnimatedContainer.tsx
  - move Skeleton.tsx → components/ui/Skeleton.tsx
  - move TechBadge.tsx → components/ui/TechBadge.tsx
  - move MorphingTitle.tsx → components/animations/MorphingTitle.tsx
  - move ProfileConfig.ts → components/config/ProfileConfig.ts
  - move DiscordProfile.tsx → components/profile/DiscordProfile.tsx
  - move Profile.tsx → components/profile/Profile.tsx

- components/animations/:
  - add index.ts

- components/config/:
  - add index.ts

- components/hooks/:
  - add index.ts

- components/profile/:
  - add index.ts

- components/ui/:
  - add index.ts
