+++ src/app/components/animations/index.ts
@@ -1,0 +1,1 @@
+export { default as MorphingTitle } from './MorphingTitle';
