+++ src/app/components/config/index.ts
@@ -1,0 +1,1 @@
+export { profileConfig } from './ProfileConfig';
