+++ src/app/components/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useDiscordSocket } from './useDiscordSocket';
+export { useDiscordPolling } from './useDiscordPolling';
