+++ src/app/components/index.ts
@@ -1,0 +1,5 @@
+export * from './ui';
+export * from './profile';
+export * from './animations';
+export * from './config';
+export * from './hooks';
