+++ src/app/components/profile/index.ts
@@ -1,0 +1,2 @@
+export { default as Profile } from './Profile';
+export { default as DiscordProfile } from './DiscordProfile';
