+++ src/app/components/ui/AnimatedContainer.tsx
