+++ src/app/components/ui/TechBadge.tsx
