+++ src/app/components/ui/index.ts
@@ -1,0 +1,3 @@
+export { default as Skeleton } from './Skeleton';
+export { default as TechBadge } from './TechBadge';
+export { default as AnimatedContainer } from './AnimatedContainer';
